add to readme and move files
